+++ packages/react-native-password/src/PasswordModal.tsx
@@ -1,10 +1,19 @@
 import React, { FC } from 'react';
 
-import { Box, Flex, helpers, Modal, NumberKeyboard, Text, Theme, useTheme, WhiteSpace } from '@td-design/react-native';
+import {
+  Box,
+  Flex,
+  helpers,
+  Modal,
+  NumberKeyboardView,
+  Text,
+  Theme,
+  useTheme,
+  WhiteSpace,
+} from '@td-design/react-native';
 
 import usePasswordModal from './usePasswordModal';
 
-const { NumberKeyboardView } = NumberKeyboard;
 const { ONE_PIXEL } = helpers;
 
 export interface PasswordModalProps {
@@ -19,13 +28,11 @@ export interface PasswordModalProps {
   /** 按下时的不透明度 */
   activeOpacity?: number;
 }
-const PasswordModal: FC<PasswordModalProps> = ({
-  length = 6,
-  activeOpacity = 0.6,
-  onDone,
-  title,
-  showCursor = false,
-}) => {
+const PasswordModal: FC<
+  PasswordModalProps & {
+    onAnimationEnd?: (visible: boolean) => void;
+  }
+> = ({ length = 6, activeOpacity = 0.6, onDone, title, showCursor = false, onAnimationEnd }) => {
   const theme = useTheme<Theme>();
   const { password, visible, setFalse, itemWidth, handleLayout, combineText, handleSubmit, handleDelete } =
     usePasswordModal({
@@ -65,7 +72,7 @@ const PasswordModal: FC<PasswordModalProps> = ({
   });
 
   return (
-    <Modal visible={visible} maskClosable={true} position="bottom" onClose={setFalse}>
+    <Modal visible={visible} maskClosable={true} position="bottom" onClose={setFalse} onAnimationEnd={onAnimationEnd}>
       <Box>
         {!!title && (
           <>