+++ packages/react-native-password/src/index.tsx
@@ -2,11 +2,20 @@ import React from 'react';
 
 import { Portal } from '@td-design/react-native';
 
-import Password from './Password';
+import { default as Password } from './Password';
 import PasswordModal, { PasswordModalProps } from './PasswordModal';
 
-function modal(props: PasswordModalProps) {
-  return Portal.add(<PasswordModal {...props} />);
+function showPasswordModal(props: PasswordModalProps) {
+  const key = Portal.add(
+    <PasswordModal
+      {...props}
+      onAnimationEnd={visible => {
+        if (!visible) {
+          Portal.remove(key);
+        }
+      }}
+    />
+  );
 }
 
-export default Object.assign(Password, { modal });
+export { Password, showPasswordModal };