+++ packages/react-native/src/form/type.ts
@@ -1,4 +1,4 @@
-import { ReactElement } from 'react';
+import { ReactElement, ReactNode } from 'react';
 
 import type { FieldProps } from 'rc-field-form/lib/Field';
 
@@ -16,7 +16,9 @@ export interface FormItemProps extends RcFieldProps {
 
 export interface FormListItemProps
   extends RcFieldProps,
-    Pick<ListItemProps, 'title' | 'required' | 'style' | 'thumb' | 'onPress' | 'arrow' | 'backgroundColor'> {
+    Pick<ListItemProps, 'required' | 'style' | 'thumb' | 'onPress' | 'arrow' | 'backgroundColor'> {
+  /** 表单项标签文本 */
+  label: ReactNode;
   /** 保持组件原始样式 */
   noStyle?: boolean;
   /** 子组件 */