+++ packages/react-native/src/number-keyboard/index.tsx
@@ -1,11 +1,22 @@
-import NumberKeyboardInput from './NumberKeyboardInput';
-import NumberKeyboardItem from './NumberKeyboardItem';
-import NumberKeyboardModal from './NumberKeyboardModal';
-import NumberKeyboardView from './NumberKeyboardView';
+import React from 'react';
 
-export default {
-  NumberKeyboardView,
-  NumberKeyboardModal,
-  NumberKeyboardItem,
-  NumberKeyboardInput,
-};
+import Portal from '../portal';
+import NumberKeyboardModal from './NumberKeyboardModal';
+import { NumberKeyboardModalProps } from './type';
+
+export function showNumberKeyboard(props: Omit<NumberKeyboardModalProps, 'visible'>) {
+  const key = Portal.add(
+    <NumberKeyboardModal
+      {...props}
+      onAnimationEnd={visible => {
+        if (!visible) {
+          Portal.remove(key);
+        }
+      }}
+    />
+  );
+}
+
+export { default as NumberKeyboardView } from './NumberKeyboardView';
+export { default as NumberKeyboardItem } from './NumberKeyboardItem';
+export { default as NumberKeyboardInput } from './NumberKeyboardInput';