+++ packages/react-native/src/progress/useCircleProgress.ts
@@ -7,18 +7,18 @@ export default function useCircleProgress({
   width = 0,
   strokeWidth = 0,
   value = 0,
-  showUnit,
-}: Pick<ProgressProps, 'width' | 'strokeWidth' | 'value' | 'showUnit'>) {
+  unit,
+}: Pick<ProgressProps, 'width' | 'strokeWidth' | 'value' | 'unit'>) {
   const radius = (width - strokeWidth) / 2;
   const circumference = 2 * Math.PI * radius;
 
-  const progress = useSharedValue(value);
-  const textLabel = useSharedValue(showUnit ? `${value}%` : `${value}`);
+  const progress = useSharedValue(0);
+  const textLabel = useSharedValue('');
 
   useEffect(() => {
     progress.value = withTiming(value, { duration: 600 });
-    textLabel.value = showUnit ? `${value}%` : `${value}`;
-  }, [value, showUnit]);
+    textLabel.value = unit ? `${value}${unit}` : `${value}`;
+  }, [value, unit]);
 
   const animatedProps = useAnimatedProps(() => ({
     strokeDashoffset: circumference - (progress.value * circumference) / 100,